+++ example/src/newFeatures/index.tsx
@@ -1,5 +1,5 @@
-import React, {memo, useCallback, useEffect, useState} from 'react';
-import {RunEnv, useAgentReducer, useMiddleWare} from "use-agent-reducer";
+import React, {memo, useCallback, useEffect} from 'react';
+import {RunEnv, useAgentReducer, useMiddleWare, useAgentMethods, useAgentSelector} from "use-agent-reducer";
 import SimpleTodoList, {SearchParamsModel} from "./model";
 import {ContentInput, PageContent, PriorLevelSelect, SearchContent} from "@/components";
 import {Button, Pagination, Table} from "antd";
@@ -12,20 +12,30 @@ import {MiddleWarePresets, weakSharing} from "agent-reducer";
 // shares state updating with each other.
 // the model created by `agent-reducer` API `weakSharing`,
 // often be reset back, if there is no living `Agent` built on it.
-const searchParamsModel = weakSharing(()=>SearchParamsModel);
+const searchParamsModel = weakSharing(() => SearchParamsModel);
 
-const simpleTodoList = weakSharing(()=>SimpleTodoList);
+const simpleTodoList = weakSharing(() => SimpleTodoList);
 
 const SearchParamComponent = memo(() => {
 
-    const {state, changeSearchContent, changeSearchPriorLevel} = useAgentReducer(searchParamsModel.current);
+    const {state, changeSearchContent, changeSearchPriorLevel, feedback} = useAgentReducer(searchParamsModel.current);
 
     // `Agent` bases on object simpleTodoList,
     // If we use class `SimpleTodoList` as a model,
     // `useAgentReducer` should create a private model object inside,
     // and then, no state updating can be shared now.
     // So, model sharing only works on 'Agents' base on a same model object.
-    const {search} = useAgentReducer(simpleTodoList.current);
+    // API `useAgentMethods` can optimize our component,
+    // it never leads its consumer (component) rerender.
+    const {search} = useAgentMethods(simpleTodoList.current);
+
+    // API `useAgentSelector` can optimize our component,
+    // it only leads its consumer (component) rerender, when the extracted data changes.
+    const searchParams = useAgentSelector(simpleTodoList.current, ({searchParams}) => searchParams);
+
+    useEffect(() => {
+        feedback(searchParams);
+    }, [searchParams]);
 
     const handleSubmit = useCallback(async () => {
         // submit current searchParams with model object `simpleTodoList`
@@ -45,8 +55,6 @@ const SearchParamComponent = memo(() => {
 
 export default function NewFeatures() {
 
-    const {feedback} = useAgentReducer(searchParamsModel.current);
-
     // `Agent` bases on model `simpleTodoList`
     const agent = useAgentReducer(simpleTodoList.current);
 
@@ -64,7 +72,6 @@ export default function NewFeatures() {
     // handle page change
     const handleChangePage = useCallback(async (currentPage: number, pageSize: number = 10) => {
         // feedback searchParams with model object `searchParamsModel`.
-        feedback(state.searchParams);
         await changePageLatest(currentPage, pageSize);
     }, [state]);