+++ docs/changes.md
@@ -2,3 +2,8 @@
 
 * [new] add API [useAgentSelector](/api?id=useagentselector) for optimizing the model sharing.
 * [new] add API [useAgentMethods](/api?id=useagentmethods) for optimizing the model sharing.
+
+## v3.2.7 2021-04-11
+
+* [update] API [useAgentSelector](/api?id=useagentselector) add a new param `equalityFn` for comparing current extracted data with the previous one, and then deside if its consumer (react component) should be rerendered.
+* [new] add API [shallowEqual](/api?id=shallowequal) for a shallow comparing with two data.